modularised the initialisation of the renderer in init.renderer()
