+++ index.html
@@ -11,7 +11,7 @@
 
     <body>
         <video id="video" width="150px" height="auto" playsinline hidden>
-            <source src="/src/videos/Tech-Vision-2019-Overview.mp4" type="video/mp4">
+            <source src="/src/assets/videos/Tech-Vision-2019-Overview.mp4" type="video/mp4">
         </video>
         <!-- include main.js file -->
         <script src="dist/main.js"></script>
