+++ src/js/store.js
@@ -1,0 +1,6 @@
+let state = {}
+
+export const getState = () => state
+export const setState = nextState => {
+    state = { ...state, ...nextState }
+}
