--- src/store.js
@@ -1,6 +1,0 @@
-let state = {}
-
-export const getState = () => state
-export const setState = nextState => {
-    state = { ...state,...nextState }
-}
