+++ src/styles/main.css
@@ -1,6 +1,6 @@
 body {
     margin: 0;
-    background-color: blue;
+    background-color: pink;
 }
 canvas {
     width: 100%;
